requirements.txt
streamlit>=1.30.0
transformers>=4.40.0
huggingface_hub>=0.18.0
torch>=2.1.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0

